+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # axeom-docs
 
+## 0.2.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @axeom/schema@0.2.0
+  - @axeom/framework@0.2.0
+  - @axeom/core@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes