+++ examples/bun/CHANGELOG.md
@@ -1,5 +1,13 @@
 # bun-example
 
+## 0.2.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @axeom/auth@0.2.0
+  - @axeom/framework@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes