+++ examples/cloudflare-workers/CHANGELOG.md
@@ -1,5 +1,13 @@
 # cloudflare-workers-example
 
+## 0.2.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @axeom/swagger@0.2.0
+  - @axeom/framework@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes