+++ examples/kitchen-sink/CHANGELOG.md
@@ -1,5 +1,22 @@
 # kitchen-sink
 
+## 0.2.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @axeom/compression@0.2.0
+  - @axeom/rate-limit@1.0.0
+  - @axeom/security@0.2.0
+  - @axeom/swagger@0.2.0
+  - @axeom/static@0.2.0
+  - @axeom/upload@0.2.0
+  - @axeom/auth@0.2.0
+  - @axeom/cors@0.2.0
+  - @axeom/ws@0.2.0
+  - @axeom/client@0.2.0
+  - @axeom/framework@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes