+++ examples/nextjs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # nextjs-example
 
+## 0.2.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @axeom/swagger@0.2.0
+  - @axeom/web@0.2.0
+  - @axeom/client@0.2.0
+  - @axeom/framework@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes