+++ packages/axeom/CHANGELOG.md
@@ -1,5 +1,24 @@
 # axeom
 
+## 0.2.0
+
+### Minor Changes
+
+- fix: standardize atomic imports and clean up meta-package exports
+
+### Patch Changes
+
+- Updated dependencies
+  - @axeom/express@0.2.0
+  - @axeom/swagger@0.2.0
+  - @axeom/logger@0.2.0
+  - @axeom/web@0.2.0
+  - @axeom/ws@0.2.0
+  - @axeom/client@0.2.0
+  - @axeom/logger-lib@0.2.0
+  - @axeom/schema@0.2.0
+  - @axeom/core@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes