+++ packages/core/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @axeom/core
 
+## 0.2.0
+
+### Minor Changes
+
+- fix: standardize atomic imports and clean up meta-package exports
+
+### Patch Changes
+
+- Updated dependencies
+  - @axeom/logger-lib@0.2.0
+
+## 0.1.1
+
+### Patch Changes
+
+- 26cf643: Initial release of Axeom framework and plugins.
+- Updated dependencies [26cf643]
+  - @axeom/logger-lib@0.1.1
+# @axeom/core
+
+## 0.2.0
+
+### Minor Changes
+
+- fix: standardize atomic imports and clean up meta-package exports
+
+### Patch Changes
+
+- Updated dependencies
+  - @axeom/logger-lib@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes