+++ packages/logger/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @axeom/logger-lib
 
+## 0.2.0
+
+### Minor Changes
+
+- fix: standardize atomic imports and clean up meta-package exports
+
 ## 0.1.1
 
 ### Patch Changes