+++ packages/plugins/swagger/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @axeom/swagger
 
+## 0.2.0
+
+### Minor Changes
+
+- fix: standardize atomic imports and clean up meta-package exports
+
+### Patch Changes
+
+- Updated dependencies
+  - @axeom/core@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes